serving/api/requirements.txt
flask
gunicorn
numpy
scikit-learn
joblib
google-cloud-storage


